feat: 중복된 태그 처리 로직 추가(#72)

+++ src/main/java/org/example/tokpik_be/tag/service/PlaceTagService.java
@@ -53,6 +53,10 @@ public class PlaceTagService {
             throw new GeneralException(TagException.INVALID_REQUEST);
         }
 
+        if(request.placeTagIds().size() != request.placeTagIds().stream().distinct().count()) {
+            throw new GeneralException(TagException.DUPLICATE_TAGS);
+        }
+
         userPlaceTagRepository.deleteByUserId(user.getId());
 
         for (long tagId : request.placeTagIds()) {
